@@ -69,7 +69,7 @@ func getAllPhishingDomains() ([]string, error) {
 	req, _ := http.NewRequest("GET", getAllPhishingDomainsUrl, nil)
 	req.Header.Add("Content-Type", "application/json")
 	req.Header.Add("Accept", "*/*")
-	req.Header.Add("X-identify", "PAGSTDB")
+	req.Header.Add("X-identify", common.ConfBotName.GetString())
 	resp, err := client.Do(req)
 	if err != nil {
 		return nil, err
@@ -89,7 +89,7 @@ func getRecentlyUpdatedPhishingDomains(seconds uint32) ([]string, []string, erro
 	req, _ := http.NewRequest("GET", reqUrl, nil)
 	req.Header.Add("Content-Type", "application/json")
 	req.Header.Add("Accept", "*/*")
-	req.Header.Add("X-identify", "PAGSTDB")
+	req.Header.Add("X-identify", common.ConfBotName.GetString())
 	resp, err := client.Do(req)
 	if err != nil {
 		return nil, nil, err
@@ -180,7 +180,7 @@ func checkRemoteForPhishingUrl(input []string) (*BitFlowAntiFishResponse, error)
 	req.Header.Add("Content-Type", "application/json")
 	req.Header.Add("Accept", "*/*")
 	req.Header.Add("Content-Length", strconv.Itoa(len(queryString)))
-	req.Header.Add("User-Agent", "PAGSTDB")
+	req.Header.Add("User-Agent", common.ConfBotUserAgent.GetString())
 
 	resp, err := client.Do(req)
 	if err != nil {
