@@ -186,16 +186,63 @@ func (i *InviteRule) Check(evt *discordgo.Message, cs *dstate.ChannelState) (del
 	return
 }
 
-type CachedInvite struct {
-	CreatedAt time.Time
-	Invite    string // we only store the invite code, no need to waste memory on the entire invite data
+type cachedInvite struct {
+	createdAt time.Time
+	guildID   int64
 }
 
-var InvitesCache = struct {
+type cachedInvites struct {
 	sync.RWMutex
+	invites map[string]cachedInvite
+}
 
-	CacheMap map[int64][]CachedInvite
-}{CacheMap: make(map[int64][]CachedInvite)}
+func (c *cachedInvites) gc(d time.Duration) {
+	ticker := time.NewTicker(d)
+	for range ticker.C {
+		c.tick(d)
+	}
+}
+
+func (c *cachedInvites) tick(d time.Duration) {
+	logger.Info("Starting invites cache GC")
+
+	t1 := time.Now()
+	var counter int
+
+	invitesCache.Lock()
+	for code, invite := range c.invites {
+		if time.Since(invite.createdAt) > d {
+			delete(c.invites, code)
+			counter++
+		}
+	}
+	invitesCache.Unlock()
+
+	logger.Infof("Finished clearing invites cache in %v. %d invites removed.", time.Since(t1), counter)
+}
+
+func (c *cachedInvites) get(code string) (cachedInvite, bool) {
+	c.RLock()
+	defer c.RUnlock()
+	i, ok := c.invites[code]
+	return i, ok
+}
+
+func (c *cachedInvites) set(code string, guildID int64) {
+	c.Lock()
+	defer c.Unlock()
+	c.invites[code] = cachedInvite{time.Now(), guildID}
+}
+
+var invitesCache cachedInvites
+
+func init() {
+	invitesCache = cachedInvites{invites: make(map[string]cachedInvite)}
+	go invitesCache.gc(invitesCacheDuration * time.Minute)
+}
+
+// invitesCacheDuration is the period between ticks for the invitesCache gc in minutes
+const invitesCacheDuration = 60
 
 func CheckMessageForBadInvites(msg string, guildID int64) (containsBadInvites bool) {
 	// check third party sites
@@ -234,27 +281,17 @@ OUTER:
 
 		checked = append(checked, id)
 
-		InvitesCache.RLock()
-		if len(InvitesCache.CacheMap) == 0 {
-			go inviteCacheGC()
-		}
-
-		for guild, cache := range InvitesCache.CacheMap {
-			for _, cachedInvite := range cache {
-				if cachedInvite.Invite == id {
-					InvitesCache.RUnlock()
-					if guild == guildID {
-						// Ignore invites to this server
-						continue OUTER
-					}
-					// If the invite is present on our cache, we return true even if it is not valid anymore
-					// This is to prevent making API Calls about invites which have a restrict rate limit
-					// Because otherwise we would most likely hit those limits
-					return true
-				}
+		i, ok := invitesCache.get(id)
+		if ok {
+			if i.guildID == guildID {
+				// Ignore invites to this server
+				continue OUTER
 			}
+
+			// If the invite is present in the cache, we return true even if it is not valid anymore.
+			// This is to prevent making API calls about invites which have a restrict rate limit.
+			return true
 		}
-		InvitesCache.RUnlock()
 
 		// Check to see if its a valid id, and if so check if its to the same server were on
 		invite, err := common.BotSession.Invite(id)
@@ -267,25 +304,7 @@ OUTER:
 			continue
 		}
 
-		InvitesCache.RLock()
-		var found bool
-		for _, cached := range InvitesCache.CacheMap[invite.Guild.ID] {
-			if cached.Invite == invite.Code {
-				found = true
-				break
-			}
-		}
-		InvitesCache.RUnlock()
-
-		if !found {
-			InvitesCache.Lock()
-			// This invite was not found on our cache, so let's add it
-			InvitesCache.CacheMap[invite.Guild.ID] = append(InvitesCache.CacheMap[invite.Guild.ID], CachedInvite{
-				CreatedAt: time.Now(),
-				Invite:    invite.Code,
-			})
-			InvitesCache.Unlock()
-		}
+		invitesCache.set(invite.Code, invite.Guild.ID)
 
 		// Ignore invites to this server
 		if invite.Guild.ID == guildID {
@@ -295,58 +314,8 @@ OUTER:
 		return true
 	}
 
-	// If we got here then there's no bad invites
+	// If we got here then there are no bad invites
 	return false
-}
-
-// Ticker duration for the InvitesCacheGC in minutes
-const InvitesCacheDuration = 60
-
-func inviteCacheGC() {
-	// Called right now, so we can sleep before starting
-	time.Sleep((InvitesCacheDuration * time.Minute) + (10 * time.Second))
-
-	ticker := time.NewTicker(InvitesCacheDuration * time.Minute)
-	for {
-		logger.Info("Starting invites cache GC")
-
-		start := time.Now()
-		var counter int
-
-		InvitesCache.Lock()
-
-		for guild, cache := range InvitesCache.CacheMap {
-			for i, cached := range cache {
-				if time.Since(cached.CreatedAt) > (InvitesCacheDuration * time.Minute) {
-					InvitesCache.CacheMap[guild] = removeFromSlice(cache, i)
-					counter++
-				}
-			}
-
-			if len(InvitesCache.CacheMap[guild]) == 0 {
-				// there are no more cached invites for this guild, so we can delete the entry
-				delete(InvitesCache.CacheMap, guild)
-			}
-		}
-
-		if len(InvitesCache.CacheMap) == 0 {
-			InvitesCache.Unlock()
-			// there are no more cached invites, so we can break the loop
-			logger.Info("No more invites cached. Breaking the Invite's GC loop now")
-			break
-		}
-
-		InvitesCache.Unlock()
-
-		logger.Infof("Finished clearing invites cache in %v. %d invites removed.", time.Since(start), counter)
-
-		<-ticker.C
-	}
-}
-
-func removeFromSlice(s []CachedInvite, i int) []CachedInvite {
-	s[i] = s[len(s)-1]
-	return s[:len(s)-1]
 }
 
 type MentionRule struct {
