@@ -147,8 +147,9 @@ func trySendWebhook(l *logrus.Entry, elem *QueuedElement) (err error) {
 	wh := whI.(*webhook)
 
 	webhookParams := &discordgo.WebhookParams{
-		Username: elem.WebhookUsername,
-		Content:  elem.MessageStr,
+		Username:  elem.WebhookUsername,
+		Content:   elem.MessageStr,
+		AvatarURL: elem.WebhookAvatarURL,
 		//AllowedMentions: &discordgo.AllowedMentions{},
 	}
 
