@@ -16,6 +16,8 @@ import (
 	"github.com/mrbentarikau/pagst/common/scheduledevents2"
 	"github.com/mrbentarikau/pagst/lib/discordgo"
 	"github.com/mrbentarikau/pagst/lib/dstate"
+
+	"golang.org/x/exp/slices"
 )
 
 var ErrTooManyCalls = errors.New("too many calls to this function")
@@ -399,9 +401,11 @@ func (c *Context) checkSafeDictNoRecursion(d Dict, n int) bool {
 }
 
 func (c *Context) tmplSendMessage(filterSpecialMentions bool, returnID bool) func(channel interface{}, msg interface{}) interface{} {
+	var repliedUser bool
 	parseMentions := []discordgo.AllowedMentionType{}
 	if !filterSpecialMentions {
 		parseMentions = append(parseMentions, discordgo.AllowedMentionTypeUsers, discordgo.AllowedMentionTypeRoles, discordgo.AllowedMentionTypeEveryone)
+		repliedUser = true
 	}
 
 	return func(channel interface{}, msg interface{}) interface{} {
@@ -423,7 +427,8 @@ func (c *Context) tmplSendMessage(filterSpecialMentions bool, returnID bool) fun
 		var m *discordgo.Message
 		msgSend := &discordgo.MessageSend{
 			AllowedMentions: discordgo.AllowedMentions{
-				Parse: parseMentions,
+				Parse:       parseMentions,
+				RepliedUser: repliedUser,
 			},
 		}
 
@@ -449,14 +454,30 @@ func (c *Context) tmplSendMessage(filterSpecialMentions bool, returnID bool) fun
 			}
 		case *discordgo.MessageSend:
 			msgSend = typedMsg
-			msgSend.AllowedMentions = discordgo.AllowedMentions{Parse: parseMentions}
+			copyAllowedMentions := typedMsg.AllowedMentions
+			msgSend.AllowedMentions = discordgo.AllowedMentions{Parse: parseMentions, RepliedUser: repliedUser}
 
-			if !filterSpecialMentions {
-				msgSend.AllowedMentions = discordgo.AllowedMentions{Parse: parseMentions}
+			if filterSpecialMentions {
+
+				if len(copyAllowedMentions.Parse) > 0 {
+					msgSend.AllowedMentions.Parse = copyAllowedMentions.Parse
+				}
+
+				if len(copyAllowedMentions.Users) > 0 &&
+					!(slices.Contains(msgSend.AllowedMentions.Parse, "users") || slices.Contains(copyAllowedMentions.Parse, "users")) {
+					msgSend.AllowedMentions.Users = copyAllowedMentions.Users
+				}
+
+				if len(copyAllowedMentions.Roles) > 0 &&
+					!(slices.Contains(msgSend.AllowedMentions.Parse, "roles") || slices.Contains(copyAllowedMentions.Parse, "roles")) {
+					msgSend.AllowedMentions.Roles = copyAllowedMentions.Roles
+				}
+
+				msgSend.AllowedMentions.RepliedUser = copyAllowedMentions.RepliedUser
 			}
 
-			if msgSend.Reference != nil {
-				cid = c.CurrentFrame.CS.ID
+			if msgSend.Reference != nil && msgSend.Reference.ChannelID == 0 {
+				//cid = c.CurrentFrame.CS.ID
 				msgSend.Reference.ChannelID = cid
 			}
 
@@ -546,7 +567,8 @@ func (c *Context) tmplEditMessage(filterSpecialMentions bool) func(channel inter
 
 		if !filterSpecialMentions {
 			msgEdit.AllowedMentions = discordgo.AllowedMentions{
-				Parse: []discordgo.AllowedMentionType{discordgo.AllowedMentionTypeUsers, discordgo.AllowedMentionTypeRoles, discordgo.AllowedMentionTypeEveryone},
+				Parse:       []discordgo.AllowedMentionType{discordgo.AllowedMentionTypeUsers, discordgo.AllowedMentionTypeRoles, discordgo.AllowedMentionTypeEveryone},
+				RepliedUser: true,
 			}
 		}
 
@@ -617,10 +639,17 @@ func (c *Context) tmplSetRoles(target interface{}, input interface{}) (string, e
 
 	rv, _ := indirect(reflect.ValueOf(input))
 	switch rv.Kind() {
+	case reflect.Int, reflect.Int64:
+		oneSlice := make([]int64, 0)
+		sliceType := reflect.TypeOf(oneSlice)
+		oneSliceReflect := reflect.MakeSlice(sliceType, 0, 0)
+		toRv := reflect.ValueOf(input)
+		oneSliceReflect = reflect.Append(oneSliceReflect, toRv)
+		rv = oneSliceReflect
 	case reflect.Slice, reflect.Array:
 		// ok
 	default:
-		return "", errors.New("value passed was not an array or slice")
+		return "", errors.New("value passed was not an array, slice or single int64")
 	}
 
 	// use a map to easily handle duplicate roles
@@ -672,7 +701,8 @@ func (c *Context) tmplSetRoles(target interface{}, input interface{}) (string, e
 		rs = append(rs, discordgo.StrID(id))
 	}
 
-	err = common.BotSession.GuildMemberEdit(c.GS.ID, targetID, rs)
+	guildMemberParams := &discordgo.GuildMemberParams{Roles: &rs}
+	_, err = common.BotSession.GuildMemberEdit(c.GS.ID, targetID, guildMemberParams)
 	if err != nil {
 		return "", err
 	}
@@ -826,9 +856,9 @@ func (c *Context) tmplDelMessage(channel, msgID interface{}, args ...interface{}
 	return ""
 }
 
-//Deletes reactions from a message either via reaction trigger or argument-set of emojis,
-//needs channelID, messageID, userID, list of emojis - up to twenty
-//can be run once per CC.
+// Deletes reactions from a message either via reaction trigger or argument-set of emojis,
+// needs channelID, messageID, userID, list of emojis - up to twenty
+// can be run once per CC.
 func (c *Context) tmplDelMessageReaction(values ...reflect.Value) (reflect.Value, error) {
 
 	f := func(args []reflect.Value) (reflect.Value, error) {
@@ -1058,6 +1088,59 @@ func (c *Context) tmplGetThread(channel interface{}) (*CtxChannel, error) {
 	return CtxChannelFromCS(cstate), nil
 }
 
+func (c *Context) tmplGetThreadsArchived(args ...interface{}) (*discordgo.ThreadsList, error) {
+	var before time.Time
+	var channelID int64
+	var limit int
+
+	if c.IncreaseCheckGenericAPICall() {
+		return nil, ErrTooManyAPICalls
+	}
+
+	if c.IncreaseCheckCallCounter("guild_archived_threads", 3) {
+		return nil, ErrTooManyCalls
+	}
+
+	if len(args) < 1 {
+		return nil, errors.New("no arguments given")
+	}
+
+	argsSdict, err := StringKeyDictionary(args...)
+	if err != nil {
+		return nil, err
+	}
+
+	for key, val := range argsSdict {
+		switch strings.ToLower(key) {
+		case "channel":
+			channelID = c.ChannelArg(val)
+			if channelID == 0 {
+				return nil, nil //dont send an error , a nil output would indicate invalid/unknown channel
+			}
+		case "before":
+			if realTime, ok := val.(time.Time); ok {
+				before = realTime
+			}
+		case "limit":
+			limit = tmplToInt(val)
+			if limit < 1 {
+				limit = 1
+			} else if limit > 100 {
+				limit = 100
+			}
+		default:
+			return nil, errors.New(`invalid key "` + key + ` "passed to getArchivedThreads builder`)
+		}
+	}
+
+	response, err := common.BotSession.ThreadsArchived(channelID, &before, limit)
+	if err != nil {
+		return nil, err
+	}
+
+	return response, nil
+}
+
 func (c *Context) tmplGetChannelOrThread(channel interface{}) (*CtxChannel, error) {
 
 	if c.IncreaseCheckGenericAPICall() {
@@ -1098,12 +1181,12 @@ func (c *Context) tmplGetChannelPins(pinCount bool) func(channel interface{}) (i
 			return len(msg), nil
 		}
 
-		ids := []int64{}
-		for _, v := range msg {
-			ids = append(ids, v.ID)
+		pinnedMessages := make([]discordgo.Message, 0, len(msg))
+		for _, m := range msg {
+			pinnedMessages = append(pinnedMessages, *m)
 		}
 
-		return ids, nil
+		return pinnedMessages, nil
 	}
 }
 
@@ -1118,7 +1201,7 @@ func (c *Context) tmplAddReactions(values ...reflect.Value) (reflect.Value, erro
 				return reflect.Value{}, ErrTooManyCalls
 			}
 
-			if err := common.BotSession.MessageReactionAdd(c.Msg.ChannelID, c.Msg.ID, reaction.String()); err != nil {
+			if err := common.BotSession.MessageReactionAdd(c.Msg.ChannelID, c.Msg.ID, fmt.Sprint(reaction)); err != nil {
 				return reflect.Value{}, err
 			}
 		}
@@ -1135,7 +1218,7 @@ func (c *Context) tmplAddResponseReactions(values ...reflect.Value) (reflect.Val
 				return reflect.Value{}, ErrTooManyCalls
 			}
 
-			c.CurrentFrame.AddResponseReactionNames = append(c.CurrentFrame.AddResponseReactionNames, reaction.String())
+			c.CurrentFrame.AddResponseReactionNames = append(c.CurrentFrame.AddResponseReactionNames, fmt.Sprint(reaction))
 		}
 		return reflect.ValueOf(""), nil
 	}
@@ -1149,7 +1232,6 @@ func (c *Context) tmplAddMessageReactions(values ...reflect.Value) (reflect.Valu
 			return reflect.Value{}, errors.New("not enough arguments (need channel and message-id)")
 		}
 
-		// cArg := args[0].Interface()
 		var cArg interface{}
 		var mID int64
 
@@ -1179,9 +1261,19 @@ func (c *Context) tmplAddMessageReactions(values ...reflect.Value) (reflect.Valu
 				return reflect.Value{}, ErrTooManyCalls
 			}
 
-			if err := common.BotSession.MessageReactionAdd(cID, mID, reaction.String()); err != nil {
+			if err := common.BotSession.MessageReactionAdd(cID, mID, fmt.Sprint(reaction)); err != nil {
 				return reflect.Value{}, err
 			}
+			/*if reaction.Kind() == reflect.String {
+				if err := common.BotSession.MessageReactionAdd(cID, mID, reaction.String()); err != nil {
+					return reflect.Value{}, err
+				}
+			} else {
+				if err := common.BotSession.MessageReactionAdd(cID, mID, fmt.Sprint(reaction.Interface())); err != nil {
+					return reflect.Value{}, err
+				}
+			}*/
+
 		}
 		return reflect.ValueOf(""), nil
 	}
@@ -1324,7 +1416,8 @@ func (c *Context) tmplEditChannelName(channel interface{}, newName string) (stri
 		return "", ErrTooManyCalls
 	}
 
-	cID := c.ChannelArgNoDMNoThread(channel)
+	//cID := c.ChannelArgNoDMNoThread(channel)
+	cID := c.ChannelArgNoDM(channel)
 	if cID == 0 {
 		return "", errors.New("unknown channel")
 	}
@@ -1333,7 +1426,9 @@ func (c *Context) tmplEditChannelName(channel interface{}, newName string) (stri
 		return "", ErrTooManyCalls
 	}
 
-	_, err := common.BotSession.ChannelEdit(cID, newName)
+	channelEdit := &discordgo.ChannelEdit{Name: newName}
+
+	_, err := common.BotSession.ChannelEdit(cID, channelEdit)
 	return "", err
 }
 
@@ -2094,7 +2189,7 @@ func (c *Context) tmplGuildMemberMove(channel interface{}, target interface{}) (
 		return "", ErrTooManyAPICalls
 	}
 
-	if c.IncreaseCheckCallCounter("guild_integrations", 5) {
+	if c.IncreaseCheckCallCounter("guild_integrations", 3) {
 		return "", ErrTooManyCalls
 	}
 
@@ -2141,6 +2236,14 @@ func (c *Context) tmplCountMembers(isBot bool) func() (int, error) {
 }
 
 func (c *Context) tmplGetAuditLog(args ...interface{}) ([]*discordgo.AuditLogEntry, error) {
+	if c.IncreaseCheckGenericAPICall() {
+		return nil, ErrTooManyAPICalls
+	}
+
+	if c.IncreaseCheckCallCounter("guild_auditlogs", 2) {
+		return nil, ErrTooManyCalls
+	}
+
 	var userID, beforeID int64
 	var actionType, limit int
 
@@ -2179,4 +2282,27 @@ func (c *Context) tmplGetAuditLog(args ...interface{}) ([]*discordgo.AuditLogEnt
 	}
 
 	return response.AuditLogEntries, nil
+}
+
+func (c *Context) tmplGetUser(target interface{}) (*discordgo.User, error) {
+	if c.IncreaseCheckGenericAPICall() {
+		return nil, ErrTooManyAPICalls
+	}
+
+	if c.IncreaseCheckCallCounter("guild_integrations", 3) {
+		return nil, ErrTooManyCalls
+	}
+
+	uID := targetUserID(target)
+	if uID == 0 {
+		return nil, nil
+	}
+
+	user, err := common.BotSession.User(uID)
+	if err != nil {
+		return nil, err
+	}
+
+	return user, nil
+
 }
