@@ -11,15 +11,15 @@ import (
 type MockBot struct {
 	SessionEstablishedFunc func(info node.SessionInfo)
 
-	StopShardFunc   func(shard int) (sessionID string, sequence int64)
-	ResumeShardFunc func(shard int, sessionID string, sequence int64)
+	StopShardFunc   func(shard int) (sessionID string, sequence int64, resumeGatewayUrl string)
+	ResumeShardFunc func(shard int, sessionID string, sequence int64, resumeGatewayUrl string)
 	AddNewShardFunc func(shards ...int)
 
 	// Caled when the bot should shut down, make sure to send EvtShutdown when completed
 	ShutdownFunc func()
 
-	InitializeShardTransferFromFunc func(shard int) (sessionID string, sequence int64)
-	InitializeShardTransferToFunc   func(shard int, sessionID string, sequence int64)
+	InitializeShardTransferFromFunc func(shard int) (sessionID string, sequence int64, resumeGatewayUrl string)
+	InitializeShardTransferToFunc   func(shard int, sessionID string, sequence int64, resumeGatewayUrl string)
 
 	// this should return when all user events has been sent, with the number of user events sent
 	StartShardTransferFromFunc func(shard int) (numEventsSent int)
@@ -35,11 +35,11 @@ func (mn *MockBot) SessionEstablished(info node.SessionInfo) {
 	}
 }
 
-func (mn *MockBot) StopShard(shard int) (sessionID string, sequence int64) {
+func (mn *MockBot) StopShard(shard int) (sessionID string, sequence int64, resumeGatewayUrl string) {
 	if mn.StopShardFunc != nil {
 		return mn.StopShardFunc(shard)
 	}
-	return "", 0
+	return "", 0, "wss://gateway.discord.com/"
 }
 
 func (mn *MockBot) AddNewShards(shards ...int) {
@@ -48,9 +48,9 @@ func (mn *MockBot) AddNewShards(shards ...int) {
 	}
 }
 
-func (mn *MockBot) ResumeShard(shard int, sessionID string, sequence int64) {
+func (mn *MockBot) ResumeShard(shard int, sessionID string, sequence int64, resumeGatewayUrl string) {
 	if mn.ResumeShardFunc != nil {
-		mn.ResumeShardFunc(shard, sessionID, sequence)
+		mn.ResumeShardFunc(shard, sessionID, sequence, resumeGatewayUrl)
 	}
 }
 
@@ -61,17 +61,17 @@ func (mn *MockBot) Shutdown() {
 	}
 }
 
-func (mn *MockBot) InitializeShardTransferFrom(shard int) (sessionID string, sequence int64) {
+func (mn *MockBot) InitializeShardTransferFrom(shard int) (sessionID string, sequence int64, resumeGatewayUrl string) {
 	if mn.InitializeShardTransferFromFunc != nil {
 		return mn.InitializeShardTransferFromFunc(shard)
 	}
 
-	return "", 0
+	return "", 0, "wss://gateway.discord.com"
 }
 
-func (mn *MockBot) InitializeShardTransferTo(shard int, sessionID string, sequence int64) {
+func (mn *MockBot) InitializeShardTransferTo(shard int, sessionID string, sequence int64, resumeGatewayUrl string) {
 	if mn.InitializeShardTransferToFunc != nil {
-		mn.InitializeShardTransferToFunc(shard, sessionID, sequence)
+		mn.InitializeShardTransferToFunc(shard, sessionID, sequence, resumeGatewayUrl)
 	}
 
 }
