@@ -29,7 +29,7 @@ var _ bot.BotInitHandler = (*Plugin)(nil)
 
 func (p *Plugin) BotInit() {
 	eventsystem.AddHandlerAsyncLastLegacy(p, p.handleMessageCreate, eventsystem.EventMessageCreate)
-	eventsystem.AddHandlerAsyncLastLegacy(p, p.handleMessageReactionAdd, eventsystem.EventMessageReactionAdd)
+	eventsystem.AddHandlerAsyncLastLegacy(p, p.handleInteractionCreate, eventsystem.EventInteractionCreate)
 	scheduledevents2.RegisterHandler("rsvp_update_session", int64(0), p.handleScheduledUpdate)
 }
 
@@ -294,6 +294,34 @@ func (p *Plugin) handleMessageCreate(evt *eventsystem.EventData) {
 	}
 }
 
+func createInteractionButtons() []discordgo.MessageComponent {
+	return []discordgo.MessageComponent{
+		discordgo.ActionsRow{
+			Components: []discordgo.MessageComponent{
+				discordgo.Button{
+					Label:    EmojiJoining,
+					Style:    discordgo.SuccessButton,
+					CustomID: EventAccepted,
+				}, discordgo.Button{
+					Label:    EmojiNotJoining,
+					Style:    discordgo.DangerButton,
+					CustomID: EventRejected,
+				},
+				discordgo.Button{
+					Label:    EmojiWaitlist,
+					Style:    discordgo.PrimaryButton,
+					CustomID: EventWaitlist,
+				},
+				discordgo.Button{
+					Label:    EmojiMaybe,
+					Style:    discordgo.PrimaryButton,
+					CustomID: EventUndecided,
+				},
+			},
+		},
+	}
+}
+
 func UpdateEventEmbed(m *models.RSVPSession) error {
 
 	usersToFetch := []int64{
@@ -341,9 +369,8 @@ func UpdateEventEmbed(m *models.RSVPSession) error {
 	embed.Description = timeUntilStr
 
 	embed.Fields = append(embed.Fields, &discordgo.MessageEmbedField{
-		Name: "Times",
-		Value: fmt.Sprintf("UTC: `%s`\nLook at the bottom of this message to see when the event starts in your local time.",
-			UTCTime.Format(timeFormat)),
+		Name:  "Times",
+		Value: fmt.Sprintf("<t:%d> (UTC: `%s`)", m.StartsAt.Unix(), UTCTime.Format(timeFormat)),
 	}, &discordgo.MessageEmbedField{
 		Name:  "Reactions usage",
 		Value: "React to mark you as a participant, undecided, or not joining",
@@ -440,7 +467,18 @@ func UpdateEventEmbed(m *models.RSVPSession) error {
 	}
 	embed.Fields = append(embed.Fields, undecidedField)
 
-	_, err := common.BotSession.ChannelMessageEditEmbed(m.ChannelID, m.MessageID, embed)
+	editMessage := discordgo.MessageEdit{
+		ID:      m.MessageID,
+		Channel: m.ChannelID,
+		Embeds:  []*discordgo.MessageEmbed{embed},
+	}
+
+	if m.StartsAt.Before(time.Now()) {
+		// Remove the buttons if event has started
+		editMessage.Components = []discordgo.MessageComponent{}
+	}
+
+	_, err := common.BotSession.ChannelMessageEditComplex(&editMessage)
 	return err
 }
 
@@ -563,7 +601,6 @@ func (p *Plugin) startEvent(m *models.RSVPSession) error {
 
 	p.sendReminders(m, "Event starting now!", "The event you signed up for: **"+m.Title+"** is starting now!")
 
-	common.BotSession.MessageReactionsRemoveAll(m.ChannelID, m.MessageID)
 	_, err := m.DeleteG(context.Background())
 	return err
 }
@@ -597,33 +634,43 @@ func (p *Plugin) sendReminders(m *models.RSVPSession, title, desc string) {
 
 }
 
-func (p *Plugin) handleMessageReactionAdd(evt *eventsystem.EventData) {
-	ra := evt.MessageReactionAdd()
-	if ra.UserID == common.BotUser.ID {
+func (p *Plugin) handleInteractionCreate(evt *eventsystem.EventData) {
+	ic := evt.InteractionCreate()
+	if ic.Type != discordgo.InteractionMessageComponent || ic.GuildID == 0 || ic.Member == nil || ic.Member.User.ID == common.BotUser.ID {
 		return
 	}
 
-	joining := ra.Emoji.Name == EmojiJoining
-	notJoining := ra.Emoji.Name == EmojiNotJoining
-	maybe := ra.Emoji.Name == EmojiMaybe
-	waitlist := ra.Emoji.Name == EmojiWaitlist
+	// Pong the interaction
+	err := common.BotSession.CreateInteractionResponse(ic.ID, ic.Token, &discordgo.InteractionResponse{
+		Type: discordgo.InteractionResponseDeferredMessageUpdate,
+	})
+	if err != nil {
+		return
+	}
+
+	eventResponse := ic.MessageComponentData().CustomID
+
+	joining := eventResponse == EventAccepted
+	notJoining := eventResponse == EventRejected
+	maybe := eventResponse == EventUndecided
+	waitlist := eventResponse == EventWaitlist
 	if !joining && !notJoining && !maybe && !waitlist {
 		return
 	}
 
-	m, err := models.RSVPSessions(models.RSVPSessionWhere.MessageID.EQ(ra.MessageID), qm.Load("RSVPSessionsMessageRSVPParticipants", qm.OrderBy("marked_as_participating_at asc"))).OneG(context.Background())
+	m, err := models.RSVPSessions(models.RSVPSessionWhere.MessageID.EQ(ic.Message.ID), qm.Load("RSVPSessionsMessageRSVPParticipants", qm.OrderBy("marked_as_participating_at asc"))).OneG(context.Background())
 	if err != nil {
 		if err == sql.ErrNoRows {
 			return
 		}
-		logger.WithError(err).WithField("guild", ra.GuildID).Error("failed retrieving RSVP session")
+		logger.WithError(err).WithField("guild", ic.GuildID).Error("failed retrieving RSVP session")
 		return
 	}
 
 	foundExisting := false
 	var participant *models.RSVPParticipant
 	for _, v := range m.R.RSVPSessionsMessageRSVPParticipants {
-		if v.UserID == ra.UserID {
+		if v.UserID == ic.Member.User.ID {
 			participant = v
 			foundExisting = true
 			break
@@ -633,8 +680,8 @@ func (p *Plugin) handleMessageReactionAdd(evt *eventsystem.EventData) {
 	if !foundExisting {
 		participant = &models.RSVPParticipant{
 			RSVPSessionsMessageID: m.MessageID,
-			UserID:                ra.UserID,
-			GuildID:               ra.GuildID,
+			UserID:                ic.Member.User.ID,
+			GuildID:               ic.GuildID,
 		}
 	}
 
@@ -675,24 +722,8 @@ func (p *Plugin) handleMessageReactionAdd(evt *eventsystem.EventData) {
 	}
 
 	if err != nil {
-		logger.WithError(err).WithField("guild", ra.GuildID).Error("failed updating rsvp participant")
+		logger.WithError(err).WithField("guild", ic.GuildID).Error("failed updating rsvp participant")
 	}
-
-	reactionsToRemove := []string{}
-	if !joining {
-		reactionsToRemove = append(reactionsToRemove, EmojiJoining)
-	}
-	if !notJoining {
-		reactionsToRemove = append(reactionsToRemove, EmojiNotJoining)
-	}
-	if !maybe {
-		reactionsToRemove = append(reactionsToRemove, EmojiMaybe)
-	}
-	if !waitlist {
-		reactionsToRemove = append(reactionsToRemove, EmojiWaitlist)
-	}
-
-	go removeReactions(ra.ChannelID, ra.MessageID, ra.UserID, reactionsToRemove...)
 
 	updatingSessiosMU.Lock()
 	for _, v := range updatingSessionEmbeds {
@@ -712,15 +743,6 @@ func (p *Plugin) handleMessageReactionAdd(evt *eventsystem.EventData) {
 	go s.run()
 	updatingSessiosMU.Unlock()
 
-}
-
-func removeReactions(channelID, messageID, userID int64, emojis ...string) {
-	for _, v := range emojis {
-		err := common.BotSession.MessageReactionRemove(channelID, messageID, v, userID)
-		if err != nil {
-			logger.WithError(err).Error("failed removing reaction")
-		}
-	}
 }
 
 var (
