@@ -9,6 +9,7 @@ import (
 
 	"github.com/PuerkitoBio/goquery"
 	"github.com/mrbentarikau/pagst/commands"
+	"github.com/mrbentarikau/pagst/common"
 	"github.com/mrbentarikau/pagst/lib/dcmd"
 )
 
@@ -36,11 +37,12 @@ var Command = &commands.YAGCommand{
 		if err != nil {
 			return "", err
 		}
-		req.Header.Set("User-Agent", "curlPAGST/7.65.1")
+
+		req.Header.Set("User-Agent", common.ConfBotUserAgent.GetString())
 
 		resp, err := http.DefaultClient.Do(req)
 		if err != nil {
-			return "", err
+			return "Something wonky happened getting results, bash.org is probably down", err
 		}
 
 		if resp.StatusCode != 200 {
