@@ -43,7 +43,7 @@ func dogFactFromAPI() ([]string, error) {
 		return nil, err
 	}
 
-	req.Header.Set("User-Agent", common.BotUserAgent)
+	req.Header.Set("User-Agent", common.ConfBotUserAgent.GetString())
 
 	resp, err := http.DefaultClient.Do(req)
 	if err != nil {
