@@ -19,12 +19,17 @@ var Command = &commands.YAGCommand{
 	Name:        "WouldYouRather",
 	Aliases:     []string{"wyr"},
 	Description: "Get presented with 2 options.",
+	ArgSwitches: []*dcmd.ArgDef{
+		{Name: "raw", Help: "Raw output"},
+	},
 	RunFunc: func(data *dcmd.Data) (interface{}, error) {
 
 		q1, q2, err := wouldYouRather()
 		if err != nil {
 			return nil, err
 		}
+
+		wyrDescription := fmt.Sprintf("**EITHER...**\n🇦 %s\n\n **OR...**\n🇧 %s", q1, q2)
 
 		content := &discordgo.MessageEmbed{
 			Author: &discordgo.MessageEmbedAuthor{
@@ -33,12 +38,17 @@ var Command = &commands.YAGCommand{
 				IconURL: "https://pagst.xyz/static/icons/favicon-32x32.png",
 			},
 			Color:       int(rand.Int63n(16777215)),
-			Description: fmt.Sprintf("**EITHER...**\n🇦 %s\n\n **OR...**\n🇧 %s", q1, q2),
+			Description: wyrDescription,
 			Footer: &discordgo.MessageEmbedFooter{
 				Text:    fmt.Sprintf("Requested by: %s#%s", data.Author.Username, data.Author.Discriminator),
 				IconURL: discordgo.EndpointUserAvatar(data.Author.ID, data.Author.Avatar),
 			},
 		}
+
+		if data.Switches["raw"].Value != nil && data.Switches["raw"].Value.(bool) {
+			return wyrDescription, nil
+		}
+
 		msg, err := common.BotSession.ChannelMessageSendEmbed(data.ChannelID, content)
 		if err != nil {
 			return nil, err
